Repaso Java 21/9/2026 Actualizacion

+++ repasoJava/src/main/java/repasoJava/EjercicioConcertCrafters/modelo/Evento.java
@@ -140,7 +140,7 @@ public abstract class Evento {
 	}
 	
 	
-	public abstract float calcularCosteBase();
+	public abstract double calcularCosteBase();
 	
 	
 	@Override
